fix: allow usage without openssl
The reqwest dependency was referenced with default features enabled
and no option to turn them off. That caused openssl to be pulled in
unconditionally and made it impossible to use rustls instead. This
change makes the dependency on reqwest/native-tls explicitly by adding
a default feature which enables it, and disables default features for
reqwest to prevent pulling it in when this crate is used without default
features.